requirements.txt
streamlit
streamlit_lottie
PyPDF2
google-generativeai
python-dotenv
watchdog